+++ andes/core/service.py
@@ -26,7 +26,8 @@ class BaseService:
     """
     Base class for Service.
 
-    Service is a v-provider type for holding internal and temporary values. Subclasses need to implement ``v``
+    Service is a v-provider type for holding internal and temporary values.
+    Subclasses need to implement ``v``
     as a member attribute or using a property decorator.
 
     Parameters
@@ -58,6 +59,25 @@ class BaseService:
         """
         return [self.name]
 
+    def assign_memory(self, n):
+        """
+        Assign memory for ``self.v`` and set the array to zero.
+
+        Parameters
+        ----------
+        n : int
+            Number of elements of the value array.
+            Provided by caller (Model.list2array).
+        """
+        self.v = np.zeros(n, dtype=self.vtype)
+
+    @property
+    def class_name(self):
+        """
+        Return the class name
+        """
+        return self.__class__.__name__
+
     @property
     def n(self):
         """
@@ -74,13 +94,6 @@ class BaseService:
             return len(self.v)
         else:
             return 1
-
-    @property
-    def class_name(self):
-        """
-        Return the class name
-        """
-        return self.__class__.__name__
 
     def __repr__(self):
         val_str = ''
@@ -124,10 +137,6 @@ class ConstService(BaseService):
         self.v_numeric = v_numeric
         self.v: Union[float, int, np.ndarray] = np.array([0.])
 
-    def assign_memory(self, n):
-        """Assign memory for ``self.v`` and set the array to zero."""
-        self.v = np.zeros(n, dtype=self.vtype)
-
 
 class VarService(ConstService):
     """
@@ -149,10 +158,11 @@ class VarService(ConstService):
 
     One can use `VarService` to implement this equation ::
 
-        self.VE = VarService(tex_name='V_E',
-                             info='VE',
-                             v_str='Abs(KPC*(vd + 1j*vq) + 1j*(KI + KPC*XL)*(Id + 1j*Iq))',
-                             )
+        self.VE = VarService(
+            tex_name='V_E',
+            info='VE',
+            v_str='Abs(KPC*(vd + 1j*vq) + 1j*(KI + KPC*XL)*(Id + 1j*Iq))',
+            )
 
     Warnings
     --------
@@ -477,10 +487,6 @@ class ExtService(BaseService):
         self.default = default
         self.v = np.array([0.])
 
-    def assign_memory(self, n):
-        """Assign memory for ``self.v`` and set the array to zero."""
-        self.v = np.zeros(n, dtype=self.vtype)
-
     def link_external(self, ext_model):
         """
         Method to be called by ``System`` for getting values from the external model or group.
@@ -621,7 +627,9 @@ class DeviceFinder(BaseService):
 
 class OperationService(BaseService):
     """
-    Base class for a type of Service which performs specific operations
+    Base class for a type of Service which performs specific operations.
+    OperationService may not use the `assign_memory` from `BaseService`,
+    because it can have a different size.
 
     This class cannot be used by itself.
 
@@ -910,6 +918,8 @@ class RefFlatten(OperationService):
 class NumSelect(OperationService):
     """
     Class for selecting values for optional NumParam.
+
+    NumSelect works with internal and external parameters.
 
     Notes
     -----
@@ -1276,7 +1286,7 @@ class CurrentSign(ConstService):
             elif bus == bus2:
                 out[idx] = -1
             else:
-                raise ValueError(f"bus {bus} is terminal of the line connecting {bus1} and {bus2}. "
+                raise ValueError(f"bus {bus} is not terminal of the line connecting {bus1} and {bus2}. "
                                  f"Check the data of {self.bus.owner.class_name}.{self.bus.name}")
 
         return out