+++ andes/models/renewable.py
@@ -892,18 +892,14 @@ class REECA1(REECA1Data, REECA1Model):
     """
     Renewable energy electrical control.
 
-    A message `RuntimeWarning: invalid value encountered in sqrt`
-    may show up following a bus-to-ground fault.
-    This warning can be safely ignored.
-    """
+    There are two user-defined voltages: `Vref0` and `Vref1`.
 
-    #
-    # The RuntimeWarning is issued from ``numpy.select``
-    # from ``Piecewise``.
-    # ``select`` evaluates all possible outcomes, including the
-    # ``sqrt`` for negative values.
-    # However, these ``nan`` intermediate values are discarded.
-    #
+    - The difference between the initial bus voltage and `Vref0`
+      should be within the voltage deadbands `dbd1` and `dbd2`.
+    - If `VFLAG=0`, the input to the second PI controller will
+      be `Vref1`.
+
+    """
 
     def __init__(self, system, config):
         REECA1Data.__init__(self)
@@ -1715,7 +1711,7 @@ class WTARA1Model(Model):
 
 class WTARA1(WTARA1Data, WTARA1Model):
     """
-    Wind turbine aerodynamics model.
+    Wind turbine aerodynamics model (no wind speed details).
     """
 
     def __init__(self, system, config):
@@ -1817,7 +1813,7 @@ class WTPTA1Model(Model):
                               e_str='-theta0 + LG_y'
                               )
 
-        self.Pord = ExtAlgeb(model='RenExciter', src='Pord', indexer=self.ree,
+        self.Pord = ExtState(model='RenExciter', src='Pord', indexer=self.ree,
                              )
 
         self.Pref = ExtAlgeb(model='RenExciter', src='Pref', indexer=self.ree,
@@ -2106,3 +2102,68 @@ class WTTQA1(WTTQA1Data, WTTQA1Model):
     def __init__(self, config, system):
         WTTQA1Data.__init__(self)
         WTTQA1Model.__init__(self, system, config)
+
+
+class WTARV1Data(ModelData):
+    """
+    Data for wind turbine aerodynamics with wind velocity initialization.
+    """
+    def __init__(self):
+        ModelData.__init__(self)
+
+        self.rego = IdxParam(mandatory=True,
+                             info='Renewable governor idx',
+                             )
+        self.nblade = NumParam(info='number of blades', default=3.0,
+                               )
+
+        self.ngen = NumParam(info='number of wind generator units', default=50,
+                             )
+        self.npole = NumParam(info='number of poles in generator', default=4,
+                              )
+        self.R = NumParam(info='rotor radius', default=30.0,
+                          unit='m',
+                          )
+        self.ngb = NumParam(info='gear box ratio', default=5,
+                            )
+        self.rho = NumParam(info='air density', unit='kg/m3', default=1.20,
+                            )
+
+
+class WTARV1Model(Model):
+    def __init__(self, system, config):
+        Model.__init__(self, system, config)
+        self.flags.tds = True
+        self.group = 'RenAerodynamics'
+
+        self.Sn = ExtParam(model='RenGovernor', src='Sn', indexer=self.rego,
+                           tex_name='S_n', export=False,
+                           )
+
+        self.Pe0 = ExtService(model='RenGovernor',
+                              src='Pe0',
+                              indexer=self.rego,
+                              tex_name='P_{e0}',
+                              )
+
+        self.Pmg = ExtAlgeb(model='RenGovernor',
+                            src='Pm',
+                            indexer=self.rego,
+                            )
+
+        self.theta = Algeb(tex_name=r'\theta',
+                           info='Pitch angle',
+                           unit='rad',
+                           )
+
+
+class WTARV1(WTARV1Data, WTARV1Model):
+    """
+    Wind turbine aerodynamics model with wind velocity details.
+
+    Work is in progress.
+    """
+
+    def __init__(self, system, config):
+        WTARV1Data.__init__(self)
+        WTARV1Model.__init__(self, system, config)