+++ tests/test_tds_resume.py
@@ -21,3 +21,14 @@ class TestTDSResume(unittest.TestCase):
         # check if values are correctly stored
         self.assertTrue(len(ss.dae.ts._ys[0.1]) > 0)
         self.assertTrue(len(ss.dae.ts._ys[0.5]) > 0)
+
+        # check if `get_data` works
+        data1 = ss.dae.ts.get_data((ss.GENROU.omega, ss.GENROU.v))
+        data2 = ss.dae.ts.get_data((ss.GENROU.omega, ss.GENROU.v), a=[2])
+        nt = len(ss.dae.ts.t)
+
+        self.assertEqual(data1.shape[1], 8)
+        self.assertEqual(data1.shape[0], nt)
+
+        self.assertEqual(data2.shape[1], 2)
+        self.assertEqual(data2.shape[0], nt)